refactor(archtest): ADAPTER-RETURNS-DECLARED-TYPES-01 迁 module-path-agnostic + 缩 baseline [#1639]
移规则体到非 test adapter_returns_declared_types.go：3 个近重复检测变体
(checkAdapterFile / countAdapterFileViolations{,NoReport}) 合并为单一
collectAdapterFileViolations []Diagnostic 核心（dogfood + Good/Bad fixture +
Check 同源）。TestImportPathToContractID 单测输入派生自 PlatformModulePath。
register=none（external-applicable 但 D2 推迟注册，move+register 合并 backlog）。
baseline 删该行。
